#514: merge custom toolbar api hack into the multi-editor architecture
i merged the current htmlarea based version of the htmleditorcontrol into the multieditor patch, fixing the svn. i also removed some code duplication as i went.

Please note that this is against our very own cvs practices as described in the developer docu. The one who commits new code must make sure not to overwrite existing updated code. The committer of the new code should be responsible of properly merging the patch with the existing code base. The linux kernel and many other os projects use this practice for a reason!

i also added svn:external properties so we pull in the editors into the right place